chore(plot): rename stacked_bar_source to stacked_bar
